docs: Move mapSpot
